Fixing ping/sst-client bugs. Converting to Slf4j logger.

+++ src/main/java/co/charbox/client/sst/utils/DataReceiver.java
@@ -4,10 +4,10 @@ package co.charbox.client.sst.utils;
 public class DataReceiver implements Runnable {
 
 	private MyIOHAndler io;
-	private int size;
+	private long size;
 	private int duration;
 	
-	public DataReceiver(MyIOHAndler io, int size) {
+	public DataReceiver(MyIOHAndler io, long size) {
 		super();
 		this.io = io;
 		this.size = size;
